\docType{data}
\name{haz_fl_occ}
\alias{haz_fl_occ}
\title{Building occupancy classes, specific to flood.}
\format{Data frame with 5 columns and 33 rows}
\description{
  Modified from Table 3.1 from HAZUS-MH MR4 Flood Model
  Technical Manual.
}
\details{
  Variables:

  \itemize{ \item Occupy_Class - Generalized occupancy
  class, e.g., RES1 and RES2 occupanices were assigned to
  the RES class. One of - AGR, AUTO, COM, EDU, GOV, IND,
  Other_Occupy, REL and RES \item Occupancy - Subclasses
  within each Occupy_Class \item Occ_Desc1 - Description of
  Occupy_Class \item Occ_Desc2 - Description of Occupancy
  provided by HAZUS \item SIC_code - SIC (Standard
  Industrial Classification) code specified by HAZUS }
}
\author{
  Gopi Goteti
}
\references{
  HAZUS-MH MR4 Flood Model User Manual and Technical
  Manual, August 2009,
  \url{http://www.fema.gov/protecting-our-communities/hazus/hazus-user-technical-manuals}.
}
\keyword{datasets}

