\docType{package}
\name{hazus}
\alias{hazus}
\alias{hazus-package}
\title{Damage functions and other utilities from FEMA's HAZUS software for
use in modeling financial risk from natural disasters.}
\description{
  Damage functions and other utilities from FEMA's HAZUS
  software for use in modeling financial risk from natural
  disasters.
}
\details{
  Damage functions are useful in modeling financial risk
  from natural disasters. \pkg{hazus} provides the damage
  functions used by FEMA's HAZUS software. \pkg{hazus}
  currently provides functionality to extract and visualize
  damage functions specific to flood hazard. Hurricane and
  earthquake damage functions would be included in the
  future.
}
\author{
  Gopi Goteti
}
\references{
  HAZUS, The Federal Emergency Management Agency's (FEMA's)
  Methodology for Estimating Potential Losses from
  Disasters, \url{http://www.fema.gov/hazus}
}

